requirements.txt
# Azure SDK and AI Services
azure-identity                                # Azure authentication and credentials management
azure-ai-projects                             # Azure AI project management
azure-ai-inference[opentelemetry]             # Azure AI inference with OpenTelemetry support
azure-ai-evaluation                           # Azure AI model evaluation tools
azure-ai-contentsafety                        # Azure content safety services
azure-search-documents==11.6.0b4              # Azure Search Documents SDK
azure-ai-ml                                   # Azure Machine Learning SDK                

# Telemetry and Monitoring
azure-monitor-opentelemetry                  # Azure Monitor OpenTelemetry integration
azure-core-tracing-opentelemetry             # Azure core OpenTelemetry tracing
opentelemetry-exporter-otlp                  # OpenTelemetry protocol exporter
opentelemetry-sdk                            # OpenTelemetry SDK

# Jupyter and Development Tools
jupyter                                      # Jupyter notebook interface
python-dotenv                                # Environment variable management
notebook                                     # Jupyter notebook core
ipykernel                                    # IPython kernel for Jupyter
pandas                                       # Data manipulation library
tabulate                                     # Pretty-print tabular data
nbformat                                     # Notebook format conversion
asyncio                                      # Async I/O support
typing                                       # Type hinting support

# Documentation Tools (MkDocs and plugins)
mkdocs-material                              # Material theme for MkDocs
mkdocs-render-swagger-plugin                 # Swagger/OpenAPI documentation
mkdocs-macros-plugin                         # Macros support for MkDocs
mkdocs-glightbox                             # Image lightbox enhancement
mkdocs-open-in-new-tab                       # Open links in new tabs
mkdocs-video                                 # Video embedding support
mkdocs-table-reader-plugin                   # Table reading capabilities
mkdocs-minify-plugin                         # Minification of HTML/CSS/JS
mkdocs-jupyter                               # Jupyter notebook integration
mkdocs-git-committers-plugin-2               # Git committers information
mkdocs-git-revision-date-localized-plugin    # Localized Git revision dates
mkdocs-git-revision-date-plugin              # Git revision date display
